requirements-dev.txt
-i https://pypi.org/simple

astroid==2.15.6

asyncmock==0.4.2

certifi==2023.7.22

charset-normalizer==3.2.0

colorama==0.4.6

coverage[toml]==7.3.0

dill==0.3.7

docutils==0.20.1

exceptiongroup==1.1.3

flit==3.9.0

flit-core==3.9.0

idna==3.4

iniconfig==2.0.0

isort==5.12.0

lazy-object-proxy==1.9.0

mccabe==0.7.0

mock==5.1.0

mypy==1.5.0

mypy-extensions==1.0.0

packaging==23.1

platformdirs==3.10.0

pluggy==1.2.0

pylint==2.17.5

pytest==7.4.0

pytest-asyncio==0.21.1

pytest-cov==4.1.0

pytest-mock==3.11.1

requests==2.31.0

toml==0.10.2

tomli==2.0.1

tomli-w==1.0.0

tomlkit==0.12.1

types-python-dateutil==2.8.19.14

typing-extensions==4.7.1

urllib3==2.0.4

wrapt==1.15.0

yapf==0.40.1

anyio==3.7.1

h11==0.14.0

h2==4.1.0

hpack==4.0.0

httpcore==0.17.3

httpx[http2]==0.24.1

hyperframe==6.0.1

microsoft-kiota-abstractions==0.7.1

sniffio==1.3.0

uritemplate==4.1.1

opentelemetry-api==1.19.0

opentelemetry-sdk==1.19.0